+++ src/services/constants/Constants.ts
@@ -2,3 +2,5 @@ export const STATUS_IN_PROGRESS = 'in_progress'
 export const STATUS_COMPLETED = 'terminated'
 export const STATUS_CANCELED = 'canceled'
 export const STATUS_PENDING = 'pending'
+
+export const ORIGIN_DRIVER = 'driver'